src/requirements.txt
asgiref==3.2.3
Django==3.0.2
django-cors-headers==3.2.1
django-environ==0.4.5
djangorestframework==3.11.0
psycopg2-binary==2.8.3
pytz==2019.3
sqlparse==0.3.0
drf-yasg=1.17.0
pyjwt=1.7.1
